==> 01b_age_effects_ind_roi.Rmd <==
---
title: "22q-del/dup-ind roi age effects"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
library(magrittr)
library(readxl)
library(writexl)
library(ggplot2)
library(mgcv)
library(gratia)
library(patchwork)
library(viridis)
library(tidyverse)
```

```{r readin_dat,include=F}
adj_df<-read_xlsx("/Users/jalbrzikowskime/Box/22q_dup_del_sMRI/data/20210628_aftercombat_ind_roi.xlsx")
old_df<-read_xlsx("/Users/jalbrzikowskime/Box/22q_dup_del_sMRI/data/20210628_aftercombat_data.xlsx")
total_df<-merge(adj_df,old_df,by="Scan_ID")
adj_df<-total_df
```





```{r age_group_effects,include=F}
adj_df$Group <- factor(adj_df$SUBJECT_IDENTITY)
adj_df$PTID<-factor(adj_df$PTID.x)

brain_regions<-c("lh_bankssts_thickness.combat","rh_bankssts_thickness.combat", "lh_entorhinal_thickness.combat", "rh_entorhinal_thickness.combat","lh_fusiform_thickness.combat",
"rh_fusiform_thickness.combat","lh_inferiortemporal_thickness.combat", "rh_inferiortemporal_thickness.combat","lh_insula_thickness.combat", "rh_insula_thickness.combat","lh_middletemporal_thickness.combat",  "rh_middletemporal_thickness.combat", "lh_parahippocampal_thickness.combat", "rh_parahippocampal_thickness.combat", "lh_superiortemporal_thickness.combat","rh_superiortemporal_thickness.combat","lh_temporalpole_thickness.combat","rh_temporalpole_thickness.combat","lh_transversetemporal_thickness.combat","rh_transversetemporal_thickness.combat","lh_inferiorparietal_thickness.combat","rh_inferiorparietal_thickness.combat","lh_isthmuscingulate_thickness.combat","rh_isthmuscingulate_thickness.combat","lh_postcentral_thickness.combat","rh_postcentral_thickness.combat", "lh_posteriorcingulate_thickness.combat", "rh_posteriorcingulate_thickness.combat", "lh_precuneus_thickness.combat","rh_precuneus_thickness.combat",  "lh_superiorparietal_thickness.combat","rh_superiorparietal_thickness.combat","lh_supramarginal_thickness.combat","rh_supramarginal_thickness.combat","rh_cuneus_thickness.combat", "rh_lateraloccipital_thickness.combat", "rh_lingual_thickness.combat","rh_pericalcarine_thickness.combat","rh_caudalanteriorcingulate_area.combat", "rh_caudalmiddlefrontal_area.combat", "rh_frontalpole_area.combat", "rh_medialorbitofrontal_area.combat", "rh_lateralorbitofrontal_area.combat", "rh_paracentral_area.combat", "rh_parsopercularis_area.combat", "rh_parsorbitalis_area.combat", "rh_parstriangularis_area.combat", "rh_precentral_area.combat", "rh_rostralanteriorcingulate_area.combat", "rh_rostralmiddlefrontal_area.combat", "rh_superiorfrontal_area.combat","lh_inferiorparietal_area.combat", "rh_inferiorparietal_area.combat", "lh_isthmuscingulate_area.combat", "rh_isthmuscingulate_area.combat", "lh_postcentral_area.combat","rh_postcentral_area.combat",  "lh_posteriorcingulate_area.combat", "rh_posteriorcingulate_area.combat","lh_precuneus_area.combat","rh_precuneus_area.combat","lh_cuneus_area.combat", "rh_cuneus_area.combat","lh_superiorparietal_area.combat", "rh_superiorparietal_area.combat","lh_supramarginal_area.combat","rh_supramarginal_area.combat",  "lh_lateraloccipital_area.combat","rh_lateraloccipital_area.combat", "lh_lingual_area.combat","rh_lingual_area.combat", "lh_pericalcarine_area.combat","rh_pericalcarine_area.combat")


results_age_group_brain<-matrix(NA,ncol=16,nrow=length(brain_regions))
colnames(results_age_group_brain)<-c("22qdel_group_t","22qdel_group_p","22qdup_group_t","22qdup_group_p","Control_age_F","Control_age_p","22qdel_age_F","22qdel_age_p","22qdup_age_F","22qdup_age_p","Cont_sig_change","22qdel_sig_change","22qdup_sig_change","Cont_22qdel_diff","Cont_22qdup_diff","22qdel_22qdup_diff")

for (i in c(1:length(brain_regions))){
  var=brain_regions[i]
  mod<-gam(adj_df[[var]]~s(Age,by=Group,k=4)+Group+SEX+s(PTID,bs="re",k=4)+EstimatedTotalIntraCranialVol.combat,selection=TRUE,method="REML",data=adj_df)


sum<-summary(mod,freq = T) 
results_age_group_brain[i,"22qdel_group_t"]<-sum$p.table[2,3]
results_age_group_brain[i,"22qdel_group_p"]<-ifelse(sum$p.table[2,4]==0,0.00000000000000022,sum$p.table[2,4])
results_age_group_brain[i,"22qdup_group_t"]<-sum$p.table[3,3]
results_age_group_brain[i,"22qdup_group_p"]<-ifelse(sum$p.table[3,4]==0,0.00000000000000022,sum$p.table[3,4])
results_age_group_brain[i,"Control_age_F"]<-sum$s.table[1,3]
results_age_group_brain[i,"Control_age_p"]<-ifelse(sum$s.table[1,4]==0,0.00000000000000022,sum$s.table[1,4])
results_age_group_brain[i,"22qdel_age_F"]<-sum$s.table[2,3]
results_age_group_brain[i,"22qdel_age_p"]<-ifelse(sum$s.table[2,4]==0,0.00000000000000022,sum$s.table[2,4])
results_age_group_brain[i,"22qdup_age_F"]<-sum$s.table[3,3]
results_age_group_brain[i,"22qdup_age_p"]<-ifelse(sum$s.table[3,4]==0,0.00000000000000022,sum$s.table[3,4])


#calculate derivatives to get periods of significant change for each group
derv<-derivatives(mod,term=c("s(Age):GroupCONTROL","s(Age):GroupPATIENT-DEL","s(Age):GroupPATIENT-DUP"))
sigjump_brain<-0.23
   cont=derv[derv$smooth=="s(Age):GroupCONTROL",]
   sig<-sign(cont$lower) == sign(cont$upper)
   contagelist<-cont$data[sig]
    j=1
    ranges=""
     if(length(contagelist)>0) {
        ranges<-round(contagelist[[j]],digits=1)
     }
    while (j < length(contagelist)) {
      j<-j+1
      diff<-contagelist[[j]]-contagelist[[j-1]]
      
      if (diff > sigjump_brain) {
        ranges<-paste0(ranges,"-",round(contagelist[[j-1]],digits=1),"|",round(contagelist[[j]],digits=1))
      }
      if(j==length(contagelist)){
        ranges<-paste0(ranges,"-",round(contagelist[[j]],digits=1))
      }
  
     results_age_group_brain[i,"Cont_sig_change"]<- ranges
    }
    
    del=derv[derv$smooth=="s(Age):GroupPATIENT-DEL",]
    sig<-sign(del$lower) == sign(del$upper)
    delagelist<-del$data[sig]

    j=1
    ranges=""
     if(length(delagelist)>0) {
      ranges<-round(delagelist[[j]],digits=1)
     }
    while (j < length(delagelist)) {
      j<-j+1
      diff<-delagelist[[j]]-delagelist[[j-1]]
      
      if (diff > sigjump_brain) {
        ranges<-paste0(ranges,"-",round(delagelist[[j-1]],digits=1),"|",round(delagelist[[j]],digits=1))
      }
      if(j==length(delagelist)){
        ranges<-paste0(ranges,"-",round(delagelist[[j]],digits=1))
      }
  
     results_age_group_brain[i,"22qdel_sig_change"]<- ranges
    }
  dup=derv[derv$smooth=="s(Age):GroupPATIENT-DUP",]
   sig<-sign(dup$lower) == sign(dup$upper)
   dupagelist<-dup$data[sig]
    j=1
    ranges=""
    if(length(dupagelist)>0) {
    ranges<-round(dupagelist[[j]],digits=1)
    }
    while (j < length(dupagelist)) {
      j<-j+1
      diff<-dupagelist[[j]]-dupagelist[[j-1]]
      
      if (diff > sigjump_brain) {
        ranges<-paste0(ranges,"-",round(dupagelist[[j-1]],digits=1),"|",round(dupagelist[[j]],digits=1))
      }
      if(j==length(dupagelist)){
        ranges<-paste0(ranges,"-",round(dupagelist[[j]],digits=1))
      }
  
     results_age_group_brain[i,"22qdup_sig_change"]<- ranges
    }


diff<-difference_smooths(mod,smooth="s(Age)")
diff_jump=0.51
#control - 22qdel
Cont_22qdel_diff<-diff[which(diff$level_1=="CONTROL" & diff$level_2=="PATIENT-DEL"),]
sig<-sign(Cont_22qdel_diff$lower) == sign(Cont_22qdel_diff$upper)
agelist<-round(Cont_22qdel_diff$Age[sig],digits=1)
    j=1
    ranges=""
     if(length(agelist)>0) {
        ranges<-round(agelist[[j]],digits=1)
     }
    while (j < length(agelist)) {
      j<-j+1
      diffs<-agelist[[j]]-agelist[[j-1]]
      
      if (diffs > diff_jump) {
        ranges<-paste0(ranges,"-",round(agelist[[j-1]],digits=1),"|",round(agelist[[j]],digits=1))
      }
      if(j==length(agelist)){
        ranges<-paste0(ranges,"-",round(agelist[[j]],digits=1))
      }
  
     results_age_group_brain[i,"Cont_22qdel_diff"]<- ranges
    }
    


Cont_22qdup_diff<-diff[which(diff$level_1=="CONTROL" & diff$level_2=="PATIENT-DUP"),]
sig<-sign(Cont_22qdup_diff$lower) == sign(Cont_22qdup_diff$upper)
agelist<-round(Cont_22qdup_diff$Age[sig],digits=1)
    j=1
    ranges=""
     if(length(agelist)>0) {
        ranges<-round(agelist[[j]],digits=1)
     }
    while (j < length(agelist)) {
      j<-j+1
      diffs<-agelist[[j]]-agelist[[j-1]]
      
      if (diffs > diff_jump) {
        ranges<-paste0(ranges,"-",round(agelist[[j-1]],digits=1),"|",round(agelist[[j]],digits=1))
      }
      if(j==length(agelist)){
        ranges<-paste0(ranges,"-",round(agelist[[j]],digits=1))
      }
  
     results_age_group_brain[i,"Cont_22qdup_diff"]<- ranges
    }


q22del_22qdup_diff<-diff[which(diff$level_1=="PATIENT-DEL" & diff$level_2=="PATIENT-DUP"),]
sig<-sign(q22del_22qdup_diff$lower) == sign(q22del_22qdup_diff$upper)
agelist<-round(q22del_22qdup_diff$Age[sig],digits = 1)
    j=1
     if(length(agelist)>0) {
        ranges<-round(agelist[[j]],digits=1)
     }
    while (j < length(agelist)) {
      j<-j+1
      diffs<-agelist[[j]]-agelist[[j-1]]
      
      if (diffs > diff_jump) {
        ranges<-paste0(ranges,"-",round(agelist[[j-1]],digits=1),"|",round(agelist[[j]],digits=1))
      }
      if(j==length(agelist)){
        ranges<-paste0(ranges,"-",round(agelist[[j]],digits=1))
      }
  
     results_age_group_brain[i,"22qdel_22qdup_diff"]<- ranges
    }
}

results_age_group_brain<-as.data.frame(results_age_group_brain)
results_age_group_brain$brain<-brain_regions
#stack all the p-values from the various comparisons into one vector
pvals<-stack(stack(results_age_group_brain[,c("22qdel_group_p","22qdup_group_p","Control_age_p","22qdel_age_p","22qdup_age_p")]) )
#adjust for multiple comparisons
pvals$qval<-p.adjust(pvals$values,method="fdr")
running_sum<-0
#now need to put parts of the adjusted vector (pvals$qval) back into the data frame
running_sum<-running_sum+length(results_age_group_brain$`22qdel_group_p`)
results_age_group_brain$`22qdel_group_q`=pvals$qval[1:running_sum]
results_age_group_brain$`22qdup_group_q`=pvals$qval[(running_sum+1):(running_sum+length(results_age_group_brain$`22qdup_group_p`))]
running_sum<-running_sum+length(results_age_group_brain$`22qdup_group_p`)
results_age_group_brain$Control_age_q=pvals$qval[(running_sum+1):(running_sum+length(results_age_group_brain$Control_age_p))]
running_sum<-running_sum+length(results_age_group_brain$Control_age_p)
results_age_group_brain$`22qdel_age_q`=pvals$qval[(running_sum+1):(running_sum+length(results_age_group_brain$`22qdel_age_p`))]
running_sum<-running_sum+length(results_age_group_brain$`22qdel_age_p`)
results_age_group_brain$`22qdup_age_q`=pvals$qval[(running_sum+1):(running_sum+length(results_age_group_brain$`22qdup_age_p`))]



### remove significant age periods from results where q>0.049
results_age_group_brain$Cont_sig_change<-ifelse(results_age_group_brain$Control_age_q>0.049,NA,results_age_group_brain$Cont_sig_change)

results_age_group_brain$`22qdel_sig_change`<-ifelse(results_age_group_brain$`22qdel_age_q`>0.049,NA,results_age_group_brain$`22qdel_sig_change`)

results_age_group_brain$`22qdup_sig_change`<-ifelse(results_age_group_brain$`22qdup_age_q`>0.049,NA,results_age_group_brain$`22qdup_sig_change`)


### remove diff comparisons where no age effects are significant
results_age_group_brain$all_insig<-(results_age_group_brain$Control_age_q>0.049 & results_age_group_brain$`22qdel_age_q`>0.049 & results_age_group_brain$`22qdup_age_q`>0.049)

results_age_group_brain$Cont_22qdel_diff<-ifelse(results_age_group_brain$all_insig,NA,results_age_group_brain$Cont_22qdel_diff)
results_age_group_brain$Cont_22qdup_diff<-ifelse(results_age_group_brain$all_insig,NA,results_age_group_brain$Cont_22qdup_diff)
results_age_group_brain$`22qdel_22qdup_diff`<-ifelse(results_age_group_brain$all_insig,NA,results_age_group_brain$`22qdel_22qdup_diff`)

results_age_group_brain<-results_age_group_brain %>% select("brain","22qdel_group_t","22qdel_group_p","22qdel_group_q","22qdup_group_t","22qdup_group_p" ,"22qdup_group_q","Control_age_F","Control_age_p","Control_age_q","22qdel_age_F","22qdel_age_p","22qdel_age_q","22qdup_age_F","22qdup_age_p","22qdup_age_q","Cont_sig_change","22qdel_sig_change","22qdup_sig_change","Cont_22qdel_diff","Cont_22qdup_diff","22qdel_22qdup_diff")


write_xlsx(results_age_group_brain,path='/Users/jalbrzikowskime/Box/22q_dup_del_sMRI/results/Aim1_ind_roi_age_results.xlsx')

```
