==> 03a_age_effects_CT_SA_ASD.Rmd <==


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(mgcv)
library(gratia)
library(readxl)
library(writexl)
library(patchwork)
library(cowplot)
```




```{r readin_dat,include=F}
adj_df<-read_xlsx('/Users/jalbrzikowskime/Box/22q_dup_del_sMRI/data/20210628_aftercombat_data.xlsx')
adj_df$Mean_thickness.combat<-adj_df$lh_MeanThickness_thickness.combat+adj_df$rh_MeanThickness_thickness.combat
adj_df$TotalSA_area.combat<-adj_df$lh_TotalSA_area.combat+adj_df$rh_TotalSA_area.combat

```

```{r ASD_mod}
#22q-del ASD specific
del22q<-adj_df[adj_df$SUBJECT_IDENTITY=="PATIENT-DEL",]
```


```{r recode_ASD,include=F}
#ASD was only formally assessed at baseline , so usually we just carried that diagnosis forward if they had it at baseline but sometimes it was sort of ‘iffy’ (like they met  it on ADOS but not ADI but were also really quiet, or something, then at followup they seemed very not ASD ). So if they got ASD at the majority of visits I would give that a ‘yes’ , and that just leaves q_0200 as a ? so for that one I’d say if they got it at baseline they should get the ASD dx. does that sound reasonable? we couldn’t really be consistent if we wanted to bc we assess ASD spectrum at each visit, and not ASD , so – there you go. hope that helps!
#q_0260: 1x = 0, 3x =1 1x=3 (*x refers to numbers of visits they received that dx)
#q_0001: 1x =0; 2x =1
#q_0200: 1x=0, 1x=1
#q_200 v1= 0, v2=3, v3=1
q0001_recode<-which(del22q$PTID=="q_0001")
del22q$ASD.DIAGNOS[q0001_recode]<-1

q0200_recode<-which(del22q$PTID=="q_0200")
del22q$ASD.DIAGNOS[q0200_recode]<-0
 
q0260_recode<-which(del22q$PTID=="q_0260")
del22q$ASD.DIAGNOS[q0260_recode]<-1

#check for any labeled 3
check<-table(del22q$PTID,del22q$ASD.DIAGNOS)


#columns 0 1 3
# q_0114 0 1 2
q0114_recode<-which(del22q$PTID=="q_0114")
del22q$ASD.DIAGNOS[q0114_recode]<-1

# q_0124 0 1 2
q0124_recode<-which(del22q$PTID=="q_0124")
del22q$ASD.DIAGNOS[q0124_recode]<-0

#q_0190 1 0 2
q0190_recode<-which(del22q$PTID=="q_0190")
del22q$ASD.DIAGNOS[q0190_recode]<-0


#q_0217 4 0 1
q_0217_recode<-which(del22q$PTID=="q_0217")
del22q$ASD.DIAGNOS[q_0217_recode]<-0


# q_0244 0 1 1
q_0244_recode<-which(del22q$PTID=="q_0244")
del22q$ASD.DIAGNOS[q_0244_recode]<-1

# q_0246	2	0	2
q_0246_recode<-which(del22q$PTID=="q_0246")
del22q$ASD.DIAGNOS[q_0246_recode]<-0

#q_0355 0 1 1
q_0355_recode<-which(del22q$PTID=="q_0355")
del22q$ASD.DIAGNOS[q_0355_recode]<-1

check2<-table(del22q$PTID,del22q$ASD.DIAGNOS)
del22q<-del22q[del22q$ASD.DIAGNOS!=3,]

del22q<-del22q[del22q$Age<26,]

unique<-del22q[!duplicated(del22q$PTID),]
table(unique$SIPSPSY.code,unique$ASD.DIAGNOS)

un<-unique[unique$SIPSPSY.code!=3,]
chisq.test(un$ASD.DIAGNOS,un$SIPSPSY.code)
```

```{r recode_PS,include=F}
del22q$psychosis_spect<-gsub(2,1,del22q$SIPSPSY.code)
del22q$psychosis_spect<-gsub(1,"psychosis_spect",del22q$psychosis_spect)
del22q$psychosis_spect<-gsub(0,"nopsychosis",del22q$psychosis_spect)
del22q$psychosis_spect<-gsub(3,"notassessed",del22q$psychosis_spect)
#del22q$psychosis_spect<-as.factor(del22q$psychosis_spect)
cont<-which(del22q$SUBJECT_IDENTITY=="CONTROL")
del22q$psychosis_spect[cont]<-"control"
del22q$psychosis_spect2<-del22q$psychosis_spect


#these are individauls who are sometimes psychosis spectrum and sometimes not psychosis spectrum
#we are recoding them as psychosis_spect (2) for all visits
recode<-c("q_0001","q_0005","q_0016","q_0021","q_0130","q_0234",
          "q_0238","q_0260","q_0263")
rec<-which(del22q$PTID=="q_0001")
del22q$psychosis_spect2[rec]<-"psychosis_spect"
rec<-which(del22q$PTID=="q_0005")
del22q$psychosis_spect2[rec]<-"psychosis_spect"
rec<-which(del22q$PTID=="q_0016")
del22q$psychosis_spect2[rec]<-"psychosis_spect"
rec<-which(del22q$PTID=="q_0021")
del22q$psychosis_spect2[rec]<-"psychosis_spect"
rec<-which(del22q$PTID=="q_0130")
del22q$psychosis_spect2[rec]<-"psychosis_spect"
rec<-which(del22q$PTID=="q_0234")
del22q$psychosis_spect2[rec]<-"psychosis_spect"
rec<-which(del22q$PTID=="q_0238")
del22q$psychosis_spect2[rec]<-"psychosis_spect"
rec<-which(del22q$PTID=="q_0260")
del22q$psychosis_spect2[rec]<-"psychosis_spect"
rec<-which(del22q$PTID=="q_0263")
del22q$psychosis_spect2[rec]<-"psychosis_spect"

del22q$psychosis_spect2<-as.factor(del22q$psychosis_spect2)
unique<-del22q[!duplicated(del22q$PTID),]
table(unique$psychosis_spect2,unique$ASD.DIAGNOS)
```


```{r ASD_analyses}

brain_regions<-c("Mean_thickness.combat","lh_MeanThickness_thickness.combat","rh_MeanThickness_thickness.combat",
              "lh_frontal_thickness.combat","rh_frontal_thickness.combat","lh_temporal_thickness.combat",
              "rh_temporal_thickness.combat","lh_parietal_thickness.combat","rh_parietal_thickness.combat",
              "lh_occipital_thickness.combat","rh_occipital_thickness.combat",
              "TotalSA_area.combat",
              "lh_frontal_area.combat","rh_frontal_area.combat",
              "lh_temporal_area.combat","rh_temporal_area.combat",
              "lh_parietal_area.combat","rh_parietal_area.combat",
              "lh_occipital_area.combat","rh_occipital_area.combat")

results_age_ASD<-matrix(NA,ncol=9,nrow=length(brain_regions))
colnames(results_age_ASD)<-c("ASD_group_t","ASD_group_p","NoASD_age_F","NoASD_age_p","ASD_age_F","ASD_age_p","22q_noASD_sig_change","22q_ASD_sig_change","ASD_nonASD_diff")


del22q$PTID<-as.factor(del22q$PTID)
del22q$ASD.DIAGNOS<-as.factor(del22q$ASD.DIAGNOS)

for (i in c(1:length(brain_regions))){
  var=brain_regions[i]
  mod<-gam(del22q[[var]]~s(Age,by=ASD.DIAGNOS,k=4)+
             ASD.DIAGNOS+SEX+SCANNER+EstimatedTotalIntraCranialVol.combat+s(PTID,bs="re"),selection=TRUE,method="REML",data=del22q)
    sum<-summary(mod,freq=T) 
    results_age_ASD[i,"ASD_group_t"]<-sum$p.table[2,3]
    results_age_ASD[i,"ASD_group_p"]<-ifelse(sum$p.table[2,4]==0,0.00000000000000022,sum$p.table[2,4])
    results_age_ASD[i,"NoASD_age_F"]<-sum$s.table[1,3]
    results_age_ASD[i,"NoASD_age_p"]<-ifelse(sum$s.table[1,4]==0,0.00000000000000022,sum$s.table[1,4])
    results_age_ASD[i,"ASD_age_F"]<-sum$s.table[2,3]
    results_age_ASD[i,"ASD_age_p"]<-ifelse(sum$s.table[2,4]==0,0.00000000000000022,sum$s.table[2,4])
    
    diff<-difference_smooths(mod,smooth="s(Age)")
    #print(var)
    draw(diff)
    #control - 22qdel
    sigjump=.39
    Cont_22qdel_diff<-diff[which(diff$level_1==0 & diff$level_2==1),]
    sig<-sign(Cont_22qdel_diff$lower) == sign(Cont_22qdel_diff$upper)
    agelist<-Cont_22qdel_diff$Age[sig]
    
    j=1
    ranges=""
    if(length(agelist)>0) {
        ranges<-round(agelist[[j]],digits=1)
       }
    while (j < length(agelist)) {
      j<-j+1
      diffs<-agelist[[j]]-agelist[[j-1]]
      
      if (diffs > sigjump) {
          ranges<-paste0(ranges,"-",round(agelist[[j-1]],digits=1),"|",round(agelist[[j]],digits=1))
      }
      if(j==length(agelist)){
        ranges<-paste0(ranges,"-",round(agelist[[j]],digits=1))
      }
      
      results_age_ASD[i,"ASD_nonASD_diff"]<-ranges
    }
    
    
    derv<-derivatives(mod,term=c("s(Age):ASD.DIAGNOS0","s(Age):ASD.DIAGNOS1"))
    noASD=derv[derv$smooth=="s(Age):ASD.DIAGNOS0",]
    sig<-sign(noASD$lower) == sign(noASD$upper)
    agelist<-round(noASD$data[sig],digits=1)
    sigjump=.21
    
    j=1
    ranges=""
    if(length(agelist)>0) {
        ranges<-round(agelist[[j]],digits=1)
       }
    while (j < length(agelist)) {
      j<-j+1
      diffs<-agelist[[j]]-agelist[[j-1]]
      
      if (diffs > sigjump) {
          ranges<-paste0(ranges,"-",round(agelist[[j-1]],digits=1),"|",round(agelist[[j]],digits=1))
      }
      if(j==length(agelist)){
        ranges<-paste0(ranges,"-",round(agelist[[j]],digits=1))
      }
      
      results_age_ASD[i,"22q_noASD_sig_change"]<-ranges
    }
    
    
    del=derv[derv$smooth=="s(Age):ASD.DIAGNOS1",]
    sig<-sign(del$lower) == sign(del$upper)
    agelist<-round(del$data[sig],digits=1)
    sigjump=.21
    
    j=1
    ranges=""
    if(length(agelist)>0) {
        ranges<-round(agelist[[j]],digits=1)
       }
    while (j < length(agelist)) {
      j<-j+1
      diffs<-agelist[[j]]-agelist[[j-1]]
      
      if (diffs > sigjump) {
          ranges<-paste0(ranges,"-",round(agelist[[j-1]],digits=1),"|",round(agelist[[j]],digits=1))
      }
      if(j==length(agelist)){
        ranges<-paste0(ranges,"-",round(agelist[[j]],digits=1))
      }
      
      results_age_ASD[i,"22q_ASD_sig_change"]<-ranges
    }
    
    
         try<-evaluate_smooth(mod,"s(Age)")
          
        try$selo <- try$est - try$se
        try$sehi <- try$est + try$se
    
        line_size=2
    p1<-ggplot(data = try, aes_string(x = "Age",y = "est") )+
          geom_ribbon(data = try,aes_string(x = "Age", ymin = "selo",ymax = "sehi", fill = "ASD.DIAGNOS"),alpha = .18, linetype = 0)+scale_fill_manual(values=c("blue","red"))+geom_line(data = try,aes_string(x = "Age", y = "est",color = "ASD.DIAGNOS"),size = 1)+
            scale_color_manual(values=c("blue","red"))+theme_bw()+theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank())+ theme(legend.title = element_blank())+ylab(var)
          
           del22q_noASD=derv[derv$smooth=="s(Age):ASD.DIAGNOS0",]
       del22q_yessychosis=derv[derv$smooth=="s(Age):ASD.DIAGNOS1",]
        
        tile_del22q_noASD <- ggplot(del22q_noASD) + aes_string(x = "data", y = 1,   
                                           fill = "derivative") + geom_raster(interpolate = FALSE)+scale_fill_gradient2(high="#dd4124",low="#00496f", mid="white", limits = c(min(derv$derivative),max(derv$derivative))) + xlab("")+ ylab("22qdel-noASD")+theme_bw()+
      theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank()) +theme(legend.position = "none")+ theme(axis.ticks.y = element_blank(),
            axis.text.y = element_blank())
          
      tile_del22q_ASD <- ggplot(del22q_yessychosis) + aes_string(x = "data", y = 1,   
                                           fill = "derivative") + geom_raster(interpolate = FALSE)+scale_fill_gradient2(high="#dd4124",low="#00496f", mid="white", limits = c(min(derv$derivative),max(derv$derivative))) + xlab("")+ ylab("22qdel-ASD")+theme_bw()+
      theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank())+ theme(axis.ticks.y = element_blank(),axis.text.y = element_blank())+labs(fill="Change \nper year")
    
     
      pdf_name<-sprintf('/Users/jalbrzikowskime/Box/22q_dup_del_sMRI/results/%s_ASD.pdf',var)
      pdf(file=pdf_name,width=7,height=8)
    print((p1|tile_del22q_noASD|tile_del22q_ASD)+plot_layout(nrow=4,heights = c(5,1,1,1)))+
      theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank()) 
      dev.off()
}




results_age_ASD<-as.data.frame(results_age_ASD)
results_age_ASD$brain<-brain_regions


#stack all the p-values from the various comparisons into one vector
pvals<-stack(stack(results_age_ASD[,c("ASD_group_p","NoASD_age_p","ASD_age_p")]) )
#adjust for multiple comparisons
pvals$qval<-p.adjust(pvals$values,method="fdr")
running_sum<-0

#now need to put parts of the adjusted vector (pvals$qval) back into the data frame
running_sum<-running_sum+length(results_age_ASD$`ASD_group_p`)
results_age_ASD$`ASD_group_q`=pvals$qval[1:running_sum]

results_age_ASD$NoASD_age_q=pvals$qval[(running_sum+1):(running_sum+length(results_age_ASD$NoASD_age_p))]
running_sum<-running_sum+length(results_age_ASD$NoASD_age_p)

results_age_ASD$`ASD_age_q`=pvals$qval[(running_sum+1):(running_sum+length(results_age_ASD$`ASD_age_p`))]

### remove significant age periods from results where q>0.049

results_age_ASD$`22q_noASD_sig_change`<-ifelse(results_age_ASD$`NoASD_age_q`>0.049,NA,results_age_ASD$`22q_noASD_sig_change`)

results_age_ASD$`22q_ASD_sig_change`<-ifelse(results_age_ASD$`ASD_age_q`>0.049,NA,results_age_ASD$`22q_ASD_sig_change`)


### remove diff comparisons where no age effects are significant
results_age_ASD$all_insig<-(results_age_ASD$NoASD_age_q>0.049 & results_age_ASD$`ASD_age_q`>0.049)

results_age_ASD$ASD_nonASD_diff<-ifelse(results_age_ASD$all_insig,NA,results_age_ASD$ASD_nonASD_diff)



results_age_ASD<-select(results_age_ASD,"brain","ASD_group_t","ASD_group_p","ASD_group_q","NoASD_age_F","NoASD_age_p","NoASD_age_q","ASD_age_F","ASD_age_p","ASD_age_q","22q_noASD_sig_change","22q_ASD_sig_change","ASD_nonASD_diff")


write_xlsx(results_age_ASD,path='/Users/jalbrzikowskime/Box/22q_dup_del_sMRI/results/CT_SA_age_ASD_results_ageless35.xlsx')

  for (i in c(1:length(brain_regions))){
var=brain_regions[i]
 name<-sprintf("/Users/jalbrzikowskime/Box/22q_dup_del_sMRI/results/smooths_diff_%s_CT_ASD.pdf",var)
pdf(name,height=5,width=8)
mod<-gam(del22q[[var]]~s(Age,by=ASD.DIAGNOS)+ASD.DIAGNOS+SEX+SCANNER+EstimatedTotalIntraCranialVol.combat+s(PTID,bs="re"),selection=TRUE,method="REML",data=del22q)
 
plot(mod,pages=1)
diff<-difference_smooths(mod,smooth="s(Age)")
print(draw(diff))
dev.off()
   }
```
